# order-service/openapi/openapi.json
{"openapi":"3.0.1","info":{"title":"Pizza Time Order Service"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/api/orders":{"get":{"tags":["Order"],"summary":"Get all orders","description":"Get all past orders","operationId":"getOrders","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Order"}}}}}}},"post":{"tags":["Order"],"summary":"Place a new order","description":"Place a new order in the store","operationId":"placeOrder","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PlaceOrderRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Order"}}}}}}},"/api/pizzas":{"get":{"tags":["OrderItem"],"summary":"Get all order items","description":"Get all available order items","operationId":"getOrderItems","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/OrderItem"}}}}}}}}},"components":{"schemas":{"Customer":{"required":["email","firstName","lastName"],"type":"object","properties":{"firstName":{"type":"string"},"lastName":{"type":"string"},"email":{"type":"string"}}},"OrderItem":{"required":["available","description","name","unitPrice"],"type":"object","properties":{"name":{"type":"string"},"description":{"type":"string"},"unitPrice":{"type":"number","format":"double"},"available":{"type":"boolean"}}},"PlaceOrderRequest":{"required":["customer","orderDate","orderItems"],"type":"object","properties":{"orderDate":{"type":"string"},"orderItems":{"type":"array","items":{"$ref":"#/components/schemas/OrderItem"}},"customer":{"$ref":"#/components/schemas/Customer"}}},"Order":{"required":["customer","orderDate","orderId","orderItems"],"type":"object","properties":{"orderId":{"type":"string"},"orderDate":{"type":"string"},"orderItems":{"type":"array","items":{"$ref":"#/components/schemas/OrderItem"}},"customer":{"$ref":"#/components/schemas/Customer"}}}}}}